requirements.txt
aws-cdk.core
aws-cdk.aws_autoscaling
aws-cdk.aws_ec2
aws-cdk.aws_eks
